# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface FacebookWebhookSubscriptionStatus {
  ok: boolean;
  subscribed: boolean;
}

export type AutoGammaOutboundMediaInputType = typeof AutoGammaOutboundMediaInputType[keyof typeof AutoGammaOutboundMediaInputType];


export const AutoGammaOutboundMediaInputType = {
  image: 'image',
  document: 'document',
  video: 'video',
  audio: 'audio',
} as const;

export interface AutoGammaOutboundMediaInput {
  type: AutoGammaOutboundMediaInputType;
  /**
     * Optional HTTPS URL that Live Chat can display
     * @pattern ^https://
     */
  url?: string;
  /**
     * Optional Meta media ID
     * @maxLength 255
     */
  id?: string;
  /** @maxLength 255 */
  filename?: string;
}

export interface AutoGammaOutboundMessageInput {
  /**
     * Meta phone_number_id that sent the message
     * @minLength 1
     * @maxLength 64
     */
  phoneNumberId: string;
  /**
     * The wamid returned by Meta for this send
     * @minLength 1
     * @maxLength 255
     */
  whatsappMessageId: string;
  /**
     * Recipient phone number in international format
     * @minLength 7
     * @maxLength 32
     */
  recipientPhone: string;
  /**
     * Fully rendered message text; an empty string is valid for media-only templates
     * @maxLength 10000
     */
  body: string;
  /** When Meta accepted the send; defaults to receipt time when omitted */
  sentAt?: string;
  media?: AutoGammaOutboundMediaInput;
}

export interface AutoGammaOutboundMessageResult {
  ok: boolean;
  created: boolean;
  contactId: string;
  messageId: string;
  whatsappMessageId: string;
}