# lib/api-zod/src/generated/types/autoGammaOutboundMediaInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AutoGammaOutboundMediaInputType } from './autoGammaOutboundMediaInputType';

export interface AutoGammaOutboundMediaInput {
  type: AutoGammaOutboundMediaInputType;
  /**
     * Optional HTTPS URL that Live Chat can display
     * @pattern ^https://
     */
  url?: string;
  /**
     * Optional Meta media ID
     * @maxLength 255
     */
  id?: string;
  /** @maxLength 255 */
  filename?: string;
}