# lib/api-zod/src/generated/types/autoGammaOutboundMessageInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AutoGammaOutboundMediaInput } from './autoGammaOutboundMediaInput';

export interface AutoGammaOutboundMessageInput {
  /**
     * Meta phone_number_id that sent the message
     * @minLength 1
     * @maxLength 64
     */
  phoneNumberId: string;
  /**
     * The wamid returned by Meta for this send
     * @minLength 1
     * @maxLength 255
     */
  whatsappMessageId: string;
  /**
     * Recipient phone number in international format
     * @minLength 7
     * @maxLength 32
     */
  recipientPhone: string;
  /**
     * Fully rendered message text; an empty string is valid for media-only templates
     * @maxLength 10000
     */
  body: string;
  /** When Meta accepted the send; defaults to receipt time when omitted */
  sentAt?: Date;
  media?: AutoGammaOutboundMediaInput;
}