# lib/api-zod/src/generated/types/autoGammaOutboundMessageResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AutoGammaOutboundMessageResult {
  ok: boolean;
  created: boolean;
  contactId: string;
  messageId: string;
  whatsappMessageId: string;
}